+++ doc/version_history.md
@@ -1,4 +1,11 @@
 ## 版本更新记录
+##### v6.5.4
+> - acp-spring-cloud-starter
+>   - [Add] 新增CloudTools
+>   - [Upgrade] 修改LogInfo本地IP获取逻辑
+> - test
+>   - [Upgrade] 取消静态ip配置 server.address
+>   - [Upgrade] 除 admin-server、gateway-server 以外均使用动态端口
 ##### v6.5.3
 > - [Upgrade] 升级依赖
 >   - kotlin coroutines 1.4.3